# Make Active Record's query cache an LRU