I don't know how prevalent this really is, but I heard several time about users having memory exhaustion issues caused by the query cache when dealing with long running jobs.

Overall it seems sensible for this cache not to be entirely unbounded.

Opening this for feedback.

TODO:
  - [x] Changelog
  - [x] Document the flag
  - [x] Decide if 50 is a good default

Ping @tenderlove in (the unlikely) case you'd remember the context on https://github.com/rails/rails/commit/9ce02118061fda778e592be83f7c4c3c7c75bfbf as I'm essentially reverting it.

Public complaints:
  - https://code.jjb.cc/turning-off-activerecord-query-cache-to-improve-memory-consumption-in-background-jobs
  - https://github.com/sidekiq/sidekiq/issues/3752#issuecomment-366001877
  - (I remember a bunch more but these search term return a lot of noise)