@nateberkopec 2023-05-02
Lovely! I would be happy to no longer have to cover this in my workshops and conf talks 😆 

50 feels low... but I'm unsure how to calibrate for the right number here.

IIRC there is a branch in the Batches code too that avoids QueryCache. If QueryCache is LRU, do we still want that branch in there? Maybe not?

QueryCache is already quite a hot path and not as fast as I would like... do we have a benchmark for it?