@casperisfine 2023-05-03
> there is a branch in the Batches code too that avoids QueryCache. If QueryCache is LRU, do we still want that branch in there? Maybe not?

The `in_batches` is very unlikely to make hits, and is likely to generate many queries. If you know a query likely won't hit, I think it still makes sense to bypass the cache.

 > QueryCache is already quite a hot path and not as fast as I would like... do we have a benchmark for it?

No, I haven't benchmarked yet. But yes, I should do that. But I don't see a lot of opportunities for performance improvement here.