@matthewd 2023-05-03
> QueryCache is already quite a hot path and not as fast as I would like... do we have a benchmark for it?

Do you mean cache _lookup_, or the full relation-through-cache-hit-to-result flow? I expect the latter to be pretty slow [compared to caching an already-loaded relation, say], because this is such a low level cache.

> context on https://github.com/rails/rails/commit/9ce02118061fda778e592be83f7c4c3c7c75bfbf as I'm essentially reverting it

Looks like it was the array allocation per cache lookup. Avoiding it in the no-binds case is an improvement, though still doesn't seem ideal. I don't have any better suggestions, though, short of separately LRUing each layer of the existing nested-hash structure.

---

It's a pretty trivial implementation, but still maybe worth pulling out an LRU class? I recall thinking that an LRU would be useful somewhere at some point in the past... but it might've been for this very thing 😅