@casperisfine 2023-05-03
> Looks like it was the array allocation per cache lookup.

That's what Aaron remembers. But that was a decade ago, small allocations like this aren't as costly.

> Avoiding it in the no-binds case is an improvement

The "no-binds" case is rather rare though, I'm not even sure if it's worth it.

> short of separately LRUing each layer of the existing nested-hash structure.

I guess if we used a dedicated class it would be easy to keep track the global size, but then it becomes hard to figure out which is the oldest inserted values. Either way I'm not convinced that this two-level hash is actually faster, I'll benchmark.