@casperisfine 2023-05-03
Ok, so here's a benchmark:

```ruby
# frozen_string_literal: true

begin
  require "bundler/inline"
rescue LoadError => e
  $stderr.puts "Bundler version 1.10 or later is required. Please update your Bundler"
  raise e
end

gemfile(true) do
  source "https://rubygems.org"
  case ENV["RAILS"]
  when "local"
    gem "activerecord", path: File.expand_path(".")
  when "edge"
    gem "activerecord", github: "rails/rails"
  else
    gem "activerecord", "~> 7.0.0"
  end
  gem "sqlite3"
  gem "benchmark-ips"
end

require "active_record"
require "minitest/autorun"
require "logger"

# This connection will do for database-independent bug reports.
ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Base.logger = Logger.new(nil)
ActiveRecord::Base.logger.level = Logger::INFO

conn = ActiveRecord::Base.connection
conn.class.class_eval do
  public(:cache_sql)
end

binds = [ActiveRecord::Relation::QueryAttribute.new("id", "10", ActiveRecord::Type::Integer.new)]

Benchmark.ips do |x|
  x.report("hit") do
    conn.cache_sql("SELECT 1", "SQL", binds) { 1 }
  end
end

Benchmark.ips do |x|
  x.report("miss") do
    conn.cache_sql("SELECT 0", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 1", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 2", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 3", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 4", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 5", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 6", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 7", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 8", "SQL", binds) { 1 }
    conn.cache_sql("SELECT 9", "SQL", binds) { 1 }
    conn.clear_query_cache
  end
end
```


7.0:
```
Warming up --------------------------------------
                 hit    13.794k i/100ms
Calculating -------------------------------------
                 hit    138.448k (± 2.3%) i/s -    703.494k in   5.083924s
Warming up --------------------------------------
                miss     6.286k i/100ms
Calculating -------------------------------------
                miss     62.456k (± 4.3%) i/s -    314.300k in   5.043301s
```

main:
```
Warming up --------------------------------------
                 hit    20.380k i/100ms
Calculating -------------------------------------
                 hit    205.475k (± 1.7%) i/s -      1.039M in   5.059895s
Warming up --------------------------------------
                miss    11.048k i/100ms
Calculating -------------------------------------
                miss    109.093k (± 1.1%) i/s -    552.400k in   5.064167s
```

This branch:
```
Warming up --------------------------------------
                 hit    18.114k i/100ms
Calculating -------------------------------------
                 hit    180.707k (± 1.4%) i/s -    905.700k in   5.013033s
Warming up --------------------------------------
                miss     4.866k i/100ms
Calculating -------------------------------------
                miss     48.961k (± 1.4%) i/s -    248.166k in   5.069772s
```

Observations:
  - I think most of the difference between `main` and `7.0` is the earlier bail-out in AR LogSubscriber. The logging of the `sql.activerecord` event really dwarfs the lookup.
  - We do indeed take a moderate ~10% slowdown on cache hits, but a much bigger one on misses.

I'll profile a bit to see if I can close the gap a bit.