@casperisfine 2023-05-03
Ok, so unsurprisingly the vast majority of the time is spent hashing the binds:

```
==================================
  Mode: cpu(1000)
  Samples: 175 (0.00% miss rate)
  GC: 11 (6.29%)
==================================
     TOTAL    (pct)     SAMPLES    (pct)     FRAME
       157  (89.7%)          77  (44.0%)     ActiveModel::Attribute#hash
        78  (44.6%)          76  (43.4%)     ActiveModel::Type::Value#hash
        11   (6.3%)          11   (6.3%)     (marking)
       160  (91.4%)           6   (3.4%)     Array#hash
       164  (93.7%)           3   (1.7%)     ActiveRecord::ConnectionAdapters::QueryCache#cache_sql
        89  (50.9%)           1   (0.6%)     Hash#delete
         1   (0.6%)           1   (0.6%)     Kernel#hash
```

https://bugs.ruby-lang.org/issues/18897 may make this significantly faster. I'll with Ruby 3.3