@casperisfine 2023-05-03
It helps a bit, but it's not stellar:

```
Warming up --------------------------------------
                 hit    19.248k i/100ms
Calculating -------------------------------------
                 hit    194.125k (± 1.7%) i/s -    981.648k in   5.058307s
Warming up --------------------------------------
                miss     5.617k i/100ms
Calculating -------------------------------------
                miss     55.951k (± 1.3%) i/s -    280.850k in   5.020406s
```

I can't really think of any way to improve the performance though.

That said, the `miss` number is for 10 misses + 1 clear. So while it indeed much slower, it's still in the `~500k i/s` range, so `~2us` instead of `~1us` while a hit on even the simplest query will likely save close to 1ms. So I think it's an acceptable hit, but happy to reconsider if some disagree.