@casperisfine 2023-05-03
Also thinking `query_cache_size` should probably be defined in the database config, not globally.